refactor: 폴더 구조 수정 및 파일명 수정(#15)
